Allow lockfiles to have reasons
Tags: #release-engineering-team

Summary: Allow callers to util.lock() to specify a reason that is written to their lock file. Additionally, when another process tries to access and encounters a lock, have that reason (along with the owning user) provided back to the caller so they know why they're blocked and by whom.

Test Plan: Not tested yet

Reviewers: mmodell, #release-engineering-team, thcipriani

Reviewed By: mmodell, #release-engineering-team, thcipriani

Subscribers: thcipriani, bd808, jenkins

Differential Revision: https://phabricator.wikimedia.org/D454