Flatten MedaWiki deploy into a single git repo
Tags: #release-engineering-team

Maniphest Tasks: T147478

Summary:
After the local compilation of wikiversions on the deployment master,
add the entirey of /srv/mediawiki to a unified git repository.

This doesn't yet attempt to change the transport mechanism for sync, nor
does it use much of the scap3 code, but it gets us to a single version
number for deployment.

This also overwrites the mediawiki-config repo as all php-* would be
ignored otherwise.

Fixes T147478

Reviewers: bd808, dduvall, demon, mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D429