Fix MediaWiki git ident errors
Summary:
Because we use GIT_* environmental variables to indicate the deployer as
the git commit author and the ssh_user as the git commiter those values
will override anything in any git configuration files. As a result, an
empty string for any one of those values will result in an ident (or
similar) error in git.

We ran into this because the `mwdeploy` system user on tin has no entry
for their fullname in their password file entry.

Reviewers: mmodell, #release-engineering-team, demon, dduvall

Reviewed By: mmodell, #release-engineering-team

Subscribers: fgiunchedi

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D545