Make checks serial by default
Summary:
This is an update so that checks better fit deployer's (and my) mental
model of checks. This also enables the workaround of having a check
using the command `sleep 30` delay the execution of a subsequent check
proposed in T156687.

Test Plan:
Run some deploys with checks and check the deploy log to make sure that command
checks are not happening concurrently.

Reviewers: dduvall, demon, mobrovac, mmodell, #release-engineering-team

Reviewed By: mobrovac

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D565