Improve TypeError message for LogstashFormatter
Summary:
If the `fields['msg']` that gets passed to the `LogstashFormatter`
contains `%<something>` it is interpreted as a place holder in a format
string. If the number of `fields['args']` does not match the number of
format strings in `fields['msg']` the logger explodes and we get no
logging of the actual message that was being sent through the logger.

Further, the TypeError message is simply, `TypeError: not enough
arguments for format string` -- what format string? What arguments? That
is not part of the message. This adds that information to the output of
the TypeError exception.

Reviewers: demon, dduvall, mmodell, #release-engineering-team, hashar

Reviewed By: mmodell, #release-engineering-team, hashar

Subscribers: bd808

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D575