Merge environment-specific scap.cfg
Summary:
The current incorrect behavior is for the environment-specific scap.cfg
file to override the global scap.cfg but not inherit setting from the
repo-specific config. This is incongruous with expectations.

When an environment-specific scap.cfg file is used it is meant to merge
with and overwrite both the repo-specific and global scap.cfg files.

By always reading the repo-specific scap.cfg file (if it exists) in
addition to the environment-specific scap.cfg file (if it exists), we
merge-in environment-specific settings with repo-specific settings.

Test Plan:
try to override a configuration value in an environment-specific
scap.cfg file without overriding git_repo and see if scap complains
about a missing git_repo and ensure that it applies the overridden
setting.

Reviewers: mobrovac, mmodell, demon, dduvall, #release-engineering-team, Zppix

Reviewed By: mobrovac, Zppix

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D616