Remove yield from Lock.lock __enter__
Summary:
When using a function as a contextmanager, you use yield to enter the
body of the `with` statement; however, when you have a class acting as a
contextmanager it is unneeded. Also, in this instance, the __enter__
function was not executing *at all* AFICT which is something for which I
don't have an explanation.

Reviewers: demon, dduvall, mmodell, #release-engineering-team

Reviewed By: demon, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D623