Create version subcommand, cleans up weird argparse voodoo
Summary: This weird behavior requires the -V argument to only be run as part of a subcommand, which is far from ideal. Yes this is technically a breaking change, but it's a meta thing

Test Plan: Tried my new `scap version` a bunch of times. Saw that `-V` no longer worked

Reviewers: #release-engineering-team, mmodell, thcipriani

Reviewed By: #release-engineering-team, mmodell, thcipriani

Subscribers: thcipriani, jenkins

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D601