Find wikiversions to remove
Summary:
Add some functions to find versions of MediaWiki that are on disk that
are old. It does this by checking the oldest reflog on the currently
checked-out branch in the directory. By default 2 versions of MediaWiki
should be kept at all times and the static assets that are less than 5
weeks old should be kept.

These functions are meant for use with `scap clean` the idea would be to
eventually automate the cleaning process completely (maybe with a [y/N]
prompt to ensure that we don't delete anything we shouldn't).

In `scap clean` something like:

    versions = tasks.get_wikiversions_ondisk()
    to_remove, to_keep_static = tasks.get_old_wikiversions(versions)

    for version in to_remove:
        print('Removing version %s entirely...' % version)
        remove_this_version_from_disk(version)

    for version in to_keep_static:
        print('Removing version %s's static assets...' % version)
        remove_all_but_static_from_disk(version)

can be used.

Reviewers: mmodell, dduvall, demon, #release-engineering-team, hashar, Zppix

Reviewed By: hashar, Zppix

Subscribers: hashar

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D618