Handle multiple services
Summary:
This patch handles restart or reload of multiple services as specified
in the scap.cfg file by a comma-separated string.

The need for this arose when deploying the jobrunner service. Although,
seemingly, trebuchet only restarted the `jobrunner` service, the
`jobchron` service also needs to be restarted. This is now achievable
by modifying the scap.cfg to read:

    service_name: jobrunner, jobchron

If one of these services only needs to be reloaded rather than restarted
the configuration would look like:

    service_name: jobrunner = reload, jobchron

This patch removes the `service_reload` configuration value. It
currently not used as far as I can tell and no longer makes sense for
multiple services.

Fixes T167098

Reviewers: hashar, demon, dduvall, mmodell, #release-engineering-team

Reviewed By: mmodell

Subscribers: Paladox, akosiaris

Tags: #release-engineering-team

Maniphest Tasks: T167098

Differential Revision: https://phabricator.wikimedia.org/D677