Don't update SHA1 for --service-restart or --dry-run
Summary:
The problem here is we update `.git/DEPLOY_HEAD` for every `scap deploy
[anything]` on the deployment server.

This may be unexpected. For instance, in the case where a user has
checked out the latest code, but first want to restart a service using
the `--service-restart` flag. This will update the `.git/DEPLOY_HEAD`
`git_rev` to the SHA1 of the code currently checked out on tin.

If after a restart, but before a full deployment, puppet changes a
remote configuration file and runs `scap deploy-local config_deploy` for
the repo it will build the config file in `[repo]-cache/[NEW
SHA1]/.git/config-files` which actually means a subsequent deployment
will fail when trying to checkout the new files to that directory since
it is non-empty.

Fixes T169141

Test Plan:
- Add a new commit to a deployment repo on the deployment server, but
  don't deploy it
- On deployment server: `scap deploy --service-restart`
- On targets: `scap deploy-local --repo [repo] config_deploy`
- On deployment server: ensure that the last `scap/sync/*` tag points to
  the last fully deployed commit (and that `git_rev` in
  `.git/DEPLOY_HEAD` is set to that commit
- On targets: ensure that config file is rendered in
  `[repo]-cache/revs/[last fully deployed SHA1]/.git/config-files` and
  not the current SHA1 at HEAD on the deployment server

Reviewers: dduvall, mmodell, demon, mobrovac, #release-engineering-team

Reviewed By: mobrovac

Tags: #release-engineering-team

Maniphest Tasks: T169141

Differential Revision: https://phabricator.wikimedia.org/D700