MediaWiki: only fail if 1/4 canaries fail
Summary:
There have been a lot of false-canary failures during deployment. I'd
like the scap canary feature to have more of a deployment bias than it
does as currently implmented.

We currently have 11 canaries in production, this means that 2 or more
would need to fail in order to fail a deployment. The current threshold
is 1 host failing.

Reviewers: mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: zeljkofilipin

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D725