Log message to explain canary behavior
Summary:
Failing a deployment due to a small spike in error rates on a single
machine has proven to be too conservative in practice. As a result we
now only fail a deployment if a the number of failing canaries is above
a threshold.

A canary failure that doesn't halt a deployment; however, looks
confusing for users. This adds a log message that should explain why a
deployment is continuing and assure users that this is not a bug.

Fixes T173146

Reviewers: mmodell, demon, #release-engineering-team

Reviewed By: demon, #release-engineering-team

Subscribers: jcrespo, Legoktm

Tags: #release-engineering-team

Maniphest Tasks: T173146

Differential Revision: https://phabricator.wikimedia.org/D939