Fix path to scap in tasks.py
Summary:
Shamelessly copy logic from cli.Application.get_script_path() into tasks.py. Idealllllyyyyy we'd clean this up and move it elsewhere but the only place it'd make sense right now is utils and that's overloaded enough as it is. Also ideally we'd kill most of tasks because it's basically helper functions for cli.Application but that pretend to be stateless (spoiler: they aren't).

cli & tasks are a mess and should probably go into a sub-module of their own. But that's outside the scope of what I'm trying to fix today (maybe I'll get drunk and fix it tonight?)

Follows-up D930 which I really really really don't want to revert.

Test Plan: Looked at it. We need to revert the hotfix on beta (really, we swapped to the *release* branch?) and then let this go out to see if it works

Reviewers: thcipriani, mmodell, #release-engineering-team

Reviewed By: thcipriani, mmodell, #release-engineering-team

Tags: #release-engineering-team

Revert Plan: I guess revert this & D930 and go back to having bin_dir still :\

Differential Revision: https://phabricator.wikimedia.org/D941