Allow passing extra arguments to rsync in sync_common()
Summary:
What it says on the tin. This is designed so we can sometimes pass rsync args that we usually wouldn't otherwise want.

In this case? We want `scap pull` to optionally support `--delete-excluded`...a flag we absolutely would not want to otherwise use. That's T157030 for the viewers playing along at home.

Test Plan: It should work.

Reviewers: mmodell, #release-engineering-team, thcipriani

Reviewed By: #release-engineering-team, thcipriani

Subscribers: thcipriani

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D981