Store more than ssh info in case of errors
Summary:
Currently when we get logs from a target we parse them pass them to the
default log handlers. The command-line log handers are silenced for
targets because it would be a ton of noise. Instead we save all logs to
the scap/logs/*.log files; however, in the case that a target fails when
running an ssh command we dump the output from that target to the cli.

That output used to be empty, except now when you pass -v you also are
passing -v to the underlying ssh commands. As a result the output is all
the ssh session information. This makes it look like there was some ssh
problem when, in fact, something else went wrong on the target.

Rather than try to silence the ssh output, lets just dump all the output
from remote command that failed.

Bug: T182623

Reviewers: mmodell, dduvall, demon, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: jenkins

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1030