MediaWiki Canaries: use swagger spec
Summary:
In addition to checking logstash for sudden increases in error-rate,
MediaWiki deployments should be able to query a known set of endpoints
for expected responses. Since this is exactly what
service-checker-swagger does for services given a url, host, and spec
path we can shell out to that as part of a deployment.

Bug: T136839

Reviewers: demon, mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: jenkins, Krinkle

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1031