Fix git.get_disclosable_head
Summary:
Currently get_disclosable_head is taking the value of `[git_dir]/HEAD`
and using that value ot determine the disclosable head. If that value
starts with `refs/heads/` or `refs/tags/` the disclosable head is a
substring of symbolic-ref in HEAD, combined with the remote (to create
the appropriate upstream). The merge-base of the derrived upstream and
HEAD is found by running:

    git merge-base HEAD [remote]/[remote-name]

For example, if `.git/HEAD` contained `refs/heads/master` and `git
remote` returned `origin` the derived upstream would be `origin/master`
and the disclosable head would be: `git merge-base HEAD origin/master`
this is all correct.

However, if the git repo is in a detatched head state, as is the case
for git submodules, there is no derived upstream and the disclosable
head becomes the merge-base of HEAD and the value found in `.git/HEAD`,
i.e., `git merget-base HEAD HEAD` which will always be the same as the
value of HEAD which is the value we often don't want to disclose.

Typically, a remote has a symbolic-ref pointing to the default remote
branch set in `.git/refs/remotes/[remote]`. That is, it is valid to use:

    git merge-base HEAD [remote]

And git is smart enough to dereference the symbolic-ref in
`.git/refs/remotes/origin/HEAD` to find the appropriate remote much of
the time.

Reviewers: demon, mmodell, #release-engineering-team

Reviewed By: demon, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1037