Scap plugins: Tidy up imports for py3 transition
Summary:
Python3 actually calls for using 'from . import bar' syntax for imports. This seems weird, as Python is not known for verbosity being a guiding trait. But when you think about it, it makes sense. When importing via ./.., you're making it completely explicit as to what module you're wanting to load. The one **relative** to your current file. So let's start moving there (and in fact, is what absolute_imports helps you when you get wrong).

Step 1: plugins.

While we're here, fix some lambda/list comprehension. It's not wrong as-is, but is easier to read and doesn't trigger idomatic warnings

Reviewers: mmodell, #release-engineering-team, thcipriani

Reviewed By: #release-engineering-team, thcipriani

Subscribers: thcipriani

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1041