Don't recurse submodules as part of fetch
Summary:
We rewrite submodules to point to the deployment server in the cache
prior to running `git submodules update` but after we run `git fetch`.
This is problematic because git fetch will recurse into submodules if
there has been a submodule bump in the parent repo by default. The
correct behavior is to never fetch submodules until after all submodule
urls are rewritten.

This is currently affecting ORES after the deployment server
switch-over.

Fixes: T196663

Test Plan:
in scap-vagrant, I did:

    ssh scap-target-01
    cd /srv/deployment/mockbase/deploy-cache/cache
    printf "%s\n\t%s\n\t%s\n' \
        '[submodule "mockbase"]'
        'path = mockbase'
        'url = http://k/mockbase/deploy/.git/modules/mockbase' > .gitmodules
    git submodule sync

This is the current prod situation, running a `scap deploy -l scap-target-01`
fails the same way as is noted in T196663. This code fixes that. I also tested
this code with a fresh deploy.

Reviewers: mmodell, demon, awight, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1066