Add ability to specify PHP version for scap
Summary:
Allows user to specify php version for scap via scap.cfg. By default
this is "php", but it will allow updates to the configuration in puppet
to enable PHP="hhvm Eval.Jit=1" for performance improvements to
rebuildLocalisationCahce.php.

The scap configuration in puppet should be updated after this is made
live. As part of the migration to php7 the configuration in puppet will,
again, need to be modified.

Bug: T203680

Reviewers: mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1110