Invalidate PHP7's opcache when needed
Summary:
PHP7 has its bytecode cache called opcache; it can operate in different
ways, but ideally we want it to never revalidate files once they're
parsed the first time, and to be able to only release new code quickly after
everything's synced. To this aim, an "admin site" for PHP7 is present in
the WMF production clusters to allow to selectively invalidate the
opcache.

Bug: T211964

Reviewers: O2 Release & Deployment Tools, mmodell, #release-engineering-team, thcipriani

Reviewed By: O2 Release & Deployment Tools, #release-engineering-team, thcipriani

Subscribers: thcipriani

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1135