Clear MessageBlobStore after syncing i18n data
This is necessary because MW's default behavior is to call
MessageBlobStore::clear() from LocalisationCache::recache(), but in our
setup that call only happens while building the new CDBs on the
deployment server. The shared MessageBlobStore cache is invalidated at
that point, but because the new CDB files aren't on the app servers yet,
the cache will be repopulated with stale data. So we need to clear the
cache again after the new CDB files are in place on the app servers.

Because the cache is shared, it's OK to run this command only once on
the deployment server, it doesn't need to run on every app server.
However, it does need to be run for each wiki.

Bug: T222539
Bug: T217719
Change-Id: I42b73d50baa9baf85c307bc8d94576f420c474e6