Update patch set 4
Patch Set 4:

thanks for doing this! actually we could use "deployment.eqiad.wmnet". That is currently an alias for deploy1001 which is correct but it would save us from updating it again next time we switch to deploy1002 or so.

Patch-set: 4
Reviewer: Dzahn <15@e9e9afe9-4712-486d-8885-f54b72dd1951>