Update patch set 4
Patch Set 4: Code-Review-1

Indeed the documentation should refer to "deployment server" and when a hostname needs to be refered to, for Wikimedia we should use the canonical hostname "deployment.eqiad.wmnet".  It is a DNS alias that is changed whenever we need to change the actual server (datacenter switch, host maintenance etc).

Regardless +1 on keeping the doc updated!

Patch-set: 4
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1