php7x: restart on 'scap pull'
Adds the option "--no-php-restart" to the "scap pull" command. Whenever
a deployer is staging a patch on a debug server, or a machine is being
repooled we run "scap pull". This will cause the action to check if
php-fpm needs restarting and run that as the mwdeploy user if it does.

We also use "scap pull" internally to sync to target machines via any of
the "scap sync*" commands. We intentionally add "--no-php-restart" to
scap sync commands since we will restart them as a last step.

The exception here is canaries, which we want to check as a first step
-- allow those to restart with "scap pull".

Bug: T224857
Change-Id: I330bed1fd7f3a0bdb1a7645c38f56f1219f24a24