php7x: restart php-fpm after all sync operations
Previous patches check for php-fpm restart on "scap pull", this extends
that functionality to the full cluster sync. Feature-flagged behind the
"php-fpm-restart-script" configuration variable.

This script will check all machines (except canaries, which have already
been checked as part of canary sync) in batches of 10% of the total
cluster size.

Bug: T224857
Change-Id: Id4401a3a17edbe0ff51b92cec1301d6dd0c36ba9