Fix: 'version' subcommand function should return 0, not None
The cli.Application abstraction or framework calls its main method
when the subcommand is run. The main method is meant to be overridden
by subclasses, which each implement a different subcommand. The
contract for the main method is that it should return 0 for success,
and something else for failure.

The version subcommand's main method returned an implicit None. This
meant that it was treated as a failure, which mean that scap wrote an
additional \a (audible bell) control character.

That, in turn, broke the smoke test of "scap version", which didn't
expect anything but the version and a newline.

Change-Id: Ib48c7c430711a239ea27e014dec0a5f05d852528