Fix scap sync without -w (--canary-wait-time)
Previously, using -w (or its long form --canary-wait-time) meant
it checked that a None value was in the range of 20..90 seconds. It
isn't, so running scap sync without the option would always fail.

The chosen fix is to drop the range check. After discussion with Tyler
Cipriani, we agreed that it makes sense to trust the deployer and not
enforce limits on the --canary-wait-time option values. This change
drops the check, and also modifies the unit test. I admit I don't
fully understand the logic of the unit test, but I think it's correct
to drop the not check for the value 30 here.

Additionally, there are some fixes for the unit test for scap sync's
option parsing. The unit test's comment referred to the option as -cw,
when -w is correct. Also, changed the asserts to be more correct and,
I think, more easy to understand.

This condition is always true in Python:

    assert 'seconds' or 'defined' in str(ve)

A non-empty string is a true value, so the or always returns the first
argument, and so the assert doesn't actually check. I changed the code
to test for 'seconds' being in the error message, or else 'defined',
or else failing. This avoids having two places in which the magic
string constants are used and should be a little less error prone
against future changes.

Also, refactored the code in scap sync's main function a bit to not
repated self.arguments.canary_wait_time many times. I found it hard to
read the long name.

Finally, this adds an integration test for running scap sync without
--canary-wait-time.

Change-Id: I0ba052299ea6482335ab343e84fcf897fd242c92