Change IOError to OSError for Python3 compatibility
Python 3 has changed the exception hierarchy compared to Python 2, to
reduce the eternal confusion of IOError vs OSError. In Python 3 they
are now basically the same, and only OSError should be used. This
commit changes the use of IOError in cdblib.py Reader to be OSError,
which fixes a failure to run the test suite under a Python version
that has the errors unified.

See
<https://www.python.org/dev/peps/pep-3151/#confusing-set-of-os-related-exceptions>

Change-Id: I3093c27b34f78ee45482cd76706891ebe6e0809c