refactor(scap.md): make subplot easier to edit in Emacs
Change chapter and section headings to use one or more # characters at
the beginning of a line to indicate heading level, instead of a row of
= or - signs. This makes Emacs outline view more useful, as it's
instantly obvious what is a chapter and what is a section.

Also, reduce the depth of the document hierarchy by raising some
sections to be chapters. There's little point in a hierarchy where
everything is a section of the same chapter.

Also, move the YAML metadata block to the end of the document. Emacs'
syntax highlighting for Markdown gets confused by it being at the top,
and since I edit this in Emacs, it's better for me to have it at the
end. I doubt anyone else really cares.

Change-Id: I2c779a10305cd8dc90df99caf6351f921af0197d