feat: add scap apply-patch for security patches during train
This is the implementation of the "scap apply-patches" subcommand,
which applies security patches during train.

This commit is a little long, but I couldn't figure out how to split
into smaller chunks in a useful way. scap/main.py has the actual
function changes, the rest is documentation for how the apply-patches
subcommand works, what its acceptance criteria are, and how they're
verified, using Subplot.

scap.yaml has bindings of scenario steps to functions.
scap.python has functions to implement the various scenario steps.

Basically, each binding matches a step using a pattern, and specifies
a function that implements the step. The pattern may capture parts of
the step that get passed to the function as an argument. Subplot
generates a test program from the markdown using the bindings and
functions.

See https://subplot.liw.fi/ for information about Subplot.

Change-Id: I56b84dbb6799244315ac30308a80caa27ca07e60