drop: build and run-time dependency on bash-completion
It's not necessary to have any dependency on bash-completion to build
a Debian package with support for bash-completion, or for a user with
bash-completion to take advantage of it. However, if Scap insists on
it being installed, someone who installs Scap gets bash-completion
even if they don't want it.

This change drops the dependencies on bash-completion. I've verified
that this doesn't affect the built package otherwise, using debdiff.

Change-Id: I287970fa0394dc6dd3fa318db6e13ec2c37b83b0

drop: runtime dependency on bash-completion

Change-Id: I6b64f70f4326c00875142470e9fafea1cc7e24c1