fix(debian): put back the build-dependency on bash-completion
The lintian tools wants bash-completion as a build-dependency if the
dh addon for bash-completion is used, even though it doesn't seem to
have an effect on the resulting .deb. But lintian fails if it's not
there, so we put it there. This undoes part of the previous change to
drop dependencies on bash-completion.

I failed to notice the lintian error yesterday, due to not using
lintian in my build testing yet. Mea culpa. I knew better.

There's still no need to have a run-time dependency on
bash-completion, though, so users are not forced to install it.

Change-Id: I2ced3edea6c699cca0276455b0a1026cb7fb3cac