fix(targets.py): patterns with both ! and []
Previously, a pattern with both ! and [], such as
"!test[01:02].eqiad.*", didn't work, because the method argument
"pattern" was shadowed by a for loop iteration variable of the same
name. This changes the name of the iteration variable, and adds a test
using both ! and [] so it doesn't happen again.

Problem was found by pylint, the redefined-argument-from-local
warning.

Change-Id: I78f097879c05f9c13a0c7ec0304d45751f96adf0