update-wikiversions: Use staging dir wikiversions to update php link
scap update-wikiversions performs a read/modify/write of
/srv/mediawiki-staging/wikiversions.json.  Then it updates the 'php'
symlink to point to the newest version found in
/srv/mediawiki/wikiversions.json (the file that wasn't modified).
Based on the comment for function update_branch_pointer:

"""Swap the php symlink over to the new version as well, if needed."""

it seems clear that it intended to use the updated version info.  This
commit fixes that by passing 'stage' to the call to
active_wikiversions.  This makes it read the just-updated
/srv/mediawiki-staging/wikiversions.json file.

Tested in train-dev environment.

Change-Id: I386bfe8f7418195ce67e12e37e74260b31e833df