Make sync-world work when there are no proxies
Before:

If there are no proxies defined, scap sync-world tells appservers to
run "scap pull" but it doesn't supply a list of deploy servers.  This
will cause the appservers to try to rsync from localhost which will
fail (unless the appserver is serving double duty as a deploy server).

After:

If there are no proxies defined, scap sync-world tells appservers to
pull from the list of masters.

This change is to support small dev environments (such as train-dev).

Change-Id: I40cb86351adea7f6f9907d6cc0a1c52ea59a3e2b