fix: find plugins again
This fixes a bug that was introduced by moving Scap plugins from
mediawiki-config into the Scap source tree: the plugins could only be
found if Scap was run in the right directory.

The fix is to look for plugins also in the directory where the
scap.plugins module is by doing a tiny bit of Python magic (using the
__file__ attribute of a module).

To prevent this from happening again, add an integration test that
verifies scap can find every command it should find, by running "scap
foo --help" for each of them.

Bug: T248490
Change-Id: I30169ab120f22cd76f85003aa9ea5f3191ca75e3