chore: format all code with Black, require it in the future
I'm tired of formatting code by hand and also of having so many
different formatting styles in the Scap code base. This change formats
all the Python code in Scap with Black, the popular Python code
formatting tool, and changes the scripts/check script to verify code
is formatted that way so that we keep the code formatted the same way
in the future.

It's a very large diff, touching every file, but it's all mechanical,
except for scripts/check (run black --check) and tox.ini (disable some
flak8 warnings that disagree with Black).

Change-Id: Icfa1ddd3db12709209305228b88689eb06086aaf