tests: pytest should not process our json logs
JSONOutputHandlerTest() uses the logging interface buts its records are
arbitrary json blobs which do not match a python logging record.

When running the testsuite with pytest under Python 3, pytest injects a
LogCaptureHandler as the root handler which is used to capture any logs
emitted and then attempts to format them using:

  DEFAULT_LOG_FORMAT = \
    "%(levelname)-8s %(name)s:%(filename)s:%(lineno)d %(message)s"

And obviously our JSON blobs lack a filename or lineno resulting in:

  "TypeError: %d format: a number is required, not NoneType"

In setUp() instruct our custom capturing logger to not propagate the
message it has handler. That effectively hide them from pytest
LogCaptureHandler and resolve the issue.

The issue does not appear under python 2 since pytest never declares the
PercentStyleMultiline class which relies on the above format.

Bug: T246025
Change-Id: I186ceab8db1d4a54ae5b0fabc8311439aeb19781