refactor: drop scap/sh.py, use subprocess directly
sh.py was 3600+ lines of intricate Python code, most of which
duplicates the functionality of the Python standard library module
subprocess.

This commit drops that file and replaces it with scap/runcmd.py,
which has a few functions that wrap subprocess for the specific cases
Scap actually needs. It's a lot hundred lines shorter than sh.py and,
I think, somewhat easier to understand.

Bug: T222372

Change-Id: I9e25409c296116d57908cdc6d990f39ac17c991f