test: use Debian packages for dependencies, not tox/pip
Stop using tox for running the Scap test suite, in favor of
scripts/check. Change the Blubber config to produce a Docker container
that has all the dependencies installed from Debian packages (just
like production has).

The Python ecosystem is dropping support for Python 2 and 3.5, which
are both end-of-life, but we're stuck with those in production for the
time being.

Also, Scap will hopefully become obsolete later this year, when we
move MediaWiki to run in Kubernetes, at which point its deployments
won't happen with Scap anymore. This also means it's not worth porting
Scap to Python 3, anymore, so the Docker container doesn't run the
Scap test suite with that.

Change-Id: I3a48aacf5a598ebbb2748c79915dbd19a78079fe