feat! rewrite apply-patches, drop list-patches and test-patches
This makes the apply-patches sub-command be more robust and give
better information to the user. It also drops the list-patches
sub-command, which was never very useful, and the test-patches
sub-command, which didn't work very reliably.

With these changes, we should be able to use apply-patches for real,
when running the train, and to set up a cron job to check that
security patches still apply.

Bug: T269153
Change-Id: I78bc30f04c9c5e7a45d6411382fc52a382832480