refactor: have gitcmd return Unicode string, not byte string
Previously all callers of gitcmd needed to convert the returned byte
string into a Unicode string. Change things so the conversion is done
in gitcmd.

If we ever need to get binary data out of git's stdout, we can add a
separate function for that, but only after we have a demonstrated
need.

Change-Id: Ia480c71e9a7f7ced6b4109e38c6d927514172b3f