Delete lock file on SIGTERM/SIGINT
Whenever scap holds a lock file and receives SIGTERM the lock file would
be left behind since that does not trigger Lock.__exit__().

Move the cleanup logic from __exit__() to a new clear_lock() method
which let us invoke the logic when receiving a signal.  Make scap exit
code be 128 + signal value like PHP or Java.

The test catch the system exit caused by the signal handler and would
exit the context manager. It thus tried to unlock a no more existing
file descriptor, the reason is that Lock.lock_fd was not being cleared
after the file got closed which is was a flaw in the code.

Bug: T192703
Change-Id: I4483f355d1fa9a588426b730b1bd1a9f5811eacf